Remove source resolver
